+++ ShipBattleGameExample/Assets/Luxodd.Game/PluginVersion.cs
@@ -4,6 +4,6 @@ namespace Luxodd.Game.Scripts
 {
     public static class PluginVersion
     {
-        public const string Version = "1.0.3";
+        public const string Version = "1.0.4";
     }
 }