+++ ShipBattleGameExample/Assets/Luxodd.Game/Scripts/Network/CommandHandler/SendSetUserDataRequestCommandHandler.cs
@@ -1,4 +1,5 @@
 using System;
+using Luxodd.Game.Scripts.Network.Payloads;
 #if NEWTONSOFT_JSON   
 using Newtonsoft.Json;
 #endif